% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_met.R
\name{mod_met}
\alias{mod_met}
\title{Compute model evaluation metrics}
\usage{
mod_met(obs = "obs", est = "est", metrics = c("ME", "MAE", "RMSE",
  "COR"))
}
\description{
Compute model evaluation metrics to asses the agreement between observed data and various reanalysis datasets
}
