% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract values from a raster brick object at the points location}
\usage{
extract_data(file, points)
}
\arguments{
\item{file}{the raster brick object}

\item{points}{represented by a two column matrix}
}
\description{
Return a data frame  the values of a raster brick object for the cell in which the points fall, and the time steps are stored as Date.
}
